refactor: consolidate splitMessage functionality into clientHelpers
- Moved the splitMessage function from utils.ts and messageManager.ts to a new clientHelpers module for better organization and reusability.
- Updated references to splitMessage across the codebase to utilize the new implementation, ensuring consistent message chunking behavior.
- Removed redundant splitMessage implementations to enhance code cleanliness and maintainability.
